Move out the mask component
